Enable multi-currency reserve currency launches

+++ src/pbaas/crosschainrpc.h
@@ -303,7 +303,7 @@ public:
     std::vector<int64_t> maxPreconvert;     // maximum amount of each reserve that can be pre-converted
 
     int32_t preLaunchDiscount;              // if non-zero, a ratio of the initial supply instead of a fixed number is used to calculate total preallocation
-    std::vector<std::pair<uint160, int64_t>> preLaunchCarveOuts; // pre-launch carve-out recipients, from reserve contributions, taken from reserve percentage
+    std::vector<std::pair<uint160, int32_t>> preLaunchCarveOuts; // pre-launch carve-out recipients, from reserve contributions, taken from reserve percentage
     std::vector<std::pair<uint160, int64_t>> preAllocation; // pre-allocation recipients, from pre-allocation/premine, emitted after reserve weights are set
     std::vector<int64_t> contributions;     // initial contributions
     int64_t initialFractionalSupply;        // initial supply available for all pre-launch conversions, not including pre-allocation, which will be added to this
@@ -333,7 +333,7 @@ public:
                         const std::vector<uint160> &Notaries, int32_t MinNotariesConfirm, int32_t BillingPeriod, int64_t NotaryReward,
                         int32_t StartBlock, int32_t EndBlock, std::vector<uint160> Currencies, std::vector<int32_t> Weights, 
                         std::vector<int64_t> Conversions, std::vector<int64_t> MinPreconvert, std::vector<int64_t> MaxPreconvert,
-                        int32_t PreLaunchDiscount, std::vector<std::pair<uint160, int64_t>> PreLaunchCarveOuts,
+                        int32_t PreLaunchDiscount, std::vector<std::pair<uint160, int32_t>> PreLaunchCarveOuts,
                         std::vector<std::pair<uint160, int64_t>> PreAllocation, std::vector<int64_t> Contributions, int64_t InitialFractionalSupply,
                         std::vector<int64_t> Preconverted, const std::vector<int64_t> &chainRewards, const std::vector<int64_t> &chainRewardsDecay,
                         const std::vector<int32_t> &chainHalving, const std::vector<int32_t> &chainEraEnd) :
